[BUGFIX] Do not turn workspace delete records into move records
Scenario: A live page has a page localization. In workspaces,
the page localization is deleted, creating a delete placeholder.
After that, the page is moved to a different parent page in
workspaces. The delete placeholder of the localization is then
turned into a move placeholder.
This is wrong since the delete information is lost.
The patch suppresses the change of delete placeholders to move
placeholders when moving records.
This issue is much harder to solve in v10 since v10 has to deal
with two move overlay records and v11 with one. The patch is
thus slated for v11 only.

Change-Id: I233b944b030bba0de6c22d4ec1b6cb862184ddb6
Resolves: #93268
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67400
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>