[TASK] Switch behavior of ArrayUtility::arrayDiffAssocRecursive()
Despite its name, the method `ArrayUtility::arrayDiffAssocRecursive()`
mimics the behavior of `array_diff_key()` and not of
`array_diff_assoc()`.

The method has been adjusted to act like `array_diff_assoc()`. As this is
considered being a breaking change, the behavior must be enabled
explicitly by passing a 3rd parameter `$useArrayDiffAssocBehavior` being
true. If the argument is either omitted or false, the old behavior is kept
but a deprecation warning will be thrown.

Resolves: #94137
Releases: master, 10.4
Change-Id: I3cfa9581596ca4114bce4daca461187e9b87fe99
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69143
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>