[TASK] Use multi record selection in workspace module
The multi record selection component is now also used
in the workspaces module. This increases consistency,
as custom checkbox logic could be removed.

Furthermore are the actions now shown above and
below the listing, similar to the recycler module.

Thanks to additional labels, it's now also clear what
each action is for and which records are affected.

Resolves: #94935
Releases: master
Change-Id: I65d275f79390deaf30b9468f40b34856b7aee8d4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70682
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>