[BUGFIX] Keep checked state of record relations in sync
The workspace review features the collapse/expand
functionality for record relations. An example is tt_content
which may contain several sys_file_reference relations.
They are included in the listing as well but collapsed by
default.

Since at least v10 those relations are always processed
together. This means, in case an editor checks a parent
record and executes an action, e.g. publish, the same
action is also executed for the relations. However,
the UX was previously not really helpful as this was
not visible to the editor.

Therefore, and to improve the UX in workspace review,
those relations are now always "bound together". This
means, when selecting the parent record all relations
are selected, too. The same goes the other way round.
When selecting a relation, all corresponding relations
as well as the parent record are selected.

As mentioned, this does not change any functionality as
this process was already done "under the hood". This
process is now just visualized for the editor.

Resolves: #94404
Releases: master, 10.4
Change-Id: I28d2c9e67747dde5ffdfb36d24f8cc7f36c246a8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70712
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>