[TASK] Avoid calling inline JavaScript jump()
LiveSearch, Shortcut and OpenDocs call jump()
and thus, uses some "magic" top-level variables
to store the actual URL to go to.

This is not ideal, and the ModuleMenu should resolve
this, preserve the state of the "id" parameter,
open the content frame and (optionally) select a module
which was in use.

The global functions and variables are deprecated now:
* top.currentSubScript
* top.currentModuleLoaded
* top.nextLoadModuleUrl
* jump()

Resolves: #95011
Releases: master
Change-Id: I173fcaf3a90d753b0397e367fb99792edeedba65
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70732
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>