[BUGFIX] Fix invalid json responses for empty page trees
JsonResponse class constructor needs to encode empty
arrays for [] to be returned as response, instead of an
empty string.

An empty string is invalid json and therefore causes frontend
error handlers like "Page tree error" to be triggered,
while there is no error if the list of available pages is empty
for an editor – it is just empty, no error happened.

Previously (new JsonResponse)->setPayload($array) has been used
to circumvent this bug in the JsonResponse construtor in order
to be able to send empty arrays as response.
These cases are now adapted to use the constructor, now that
it is working as expected.

Also adapt to use strict typing and remove unneeded runtime
resource checks while at it. JsonResponse is internal API,
therefore this is not considered breaking.

Releases: master
Resolves: #95484
Change-Id: Ic3cabe20aeefb97b291fdcfe7ea52e8627993b63
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71451
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>