[BUGFIX] Use correct selector for progress bar in workspaces
Since #91085, sending a remote request in workspaces,
e.g. sending records to another stage, is visualized by
a progress bar. The configured selector for the progress
bar however differs between the review module and the
preview module and therefore led to a TypeError.

This is now fixed by using the correct selector
depending on the current module.

Resolves: #95661
Related: #91085
Releases: master
Change-Id: Ibeb84879e459ac8e94a0e248d77b2d974b94710c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72080
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>