[TASK] More deterministic functional test 'regular' DataSet
The DataHandler 'regular' functional import DataSet
uses a duplicate sorting value for tt_content records
296 and 297.

This is an artificial setup - the DataHandler won't
create this itself.

To make sorting values in tests easier to understand and
follow, the import set is adapted to use distinct sorting
values for these rows. This is a pre-patch to fix a
localization related sorting issue.

In case the "duplicate sorting value" scenario should
be tested, it should receive dedicated tests.

Resolves: #96035
Related: #83079
Releases: master, 11.5
Change-Id: I281eae5ca8f57ed8e16e147d27ce7c40f07b7ffe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72234
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>