[BUGFIX] Discard 'new' workspace element when deleting live
Scenario:
* Create a page in live
* Create a default language content element in live
* Localize live page
* Select a workspace
* Localize the content element in workspaces, translate mode
* Select live workspace
* Delete default content element in live

When deleting the default language content element in live,
usually all translated content elements are deleted, too.
This does not work for content element translations that
have been created in workspaces (t3ver_state=1). Those
elements *should* be discarded, but are not.

The patch adds a method to DataHandler to take care of
discarding t3ver_state=1 translations in workspaces when
a live record is set to deleted, and adds two functional
tests to verify behavior.

Resolves: #97018
Releases: main, 11.5
Change-Id: I1bba5de52447cdfb9e80feca27aebf9855dccd28
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73659
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>