[TASK] Unify log messages in DataHandler to be PSR-3 compliant
This change adapts DataHandler logs to use
interpolation from PSR-3 instead of %s to identify
variables in log_details properly later-on.

This makes all log messages from DataHandler and Workspaces
PSR-3 compliant.

In addition, all messages do not contain a "!" at the end anymore,
as there is no need to emphasize errors through the log message
itself (that's why there is type "Error" for logging).
Same goes for "." in order to be consistent.

Resolves: #97556
Releases: main
Change-Id: Ia509b6e121be64cf89d111bdc260332d343b164c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74526
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>