[TASK] Switch to dart-sass
node-sass respectively LibSass is deprecated as stated in
https://www.npmjs.com/package/node-sass

Since we use Bootstrap v5 since TYPO3 v11, which itself compiles with
dart-sass, the native JavaScript based sass solution, we can finally
also make the switch. This is also possible, as FontAwesome is now
dropped from the core.

The sass-migrator was used here to migrate the majority of division
operators to math.div or inverse it to a multiplication.
See: https://sass-lang.com/documentation/cli/migrator

Used commands:

> yarn remove node-sass
> yarn add sass -D
> sass-migrator division **/*.scss
> Build/Scripts/runTests.sh -s buildCss

Resolves: #97538
Releases: main
Change-Id: I17557ed98d6d30421c05b4a7d934a83654606d4f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75433
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>