[TASK] Streamline spacings
General spacings have been streamlined that are most relevant for
CSS declarations that add `margin-bottom` to ensure there is space
available to the next element.

Headlines h1, h2, and h3 will not add additional space to the top if
they are used in a document flow and are not the first element, and
have no classes assigned. This provides a visual divider when a
new content section is opened while keeping existing styles untouched.

There are edge cases where it can be necessary to add the space
manually because some additional necessary nesting is in place. For
this case, the class `.headline-spaced` was introduced. This class is
limited to h1, h2, and h3 and should be used with care.

The code styling for <pre> and <code> has been adjusted when
`[class*="language-"]` is matching. While we have no automatic syntax
highlighting for this kind of area yet, this indicates that a user
wants to highlight a section of code.

Dropping row and col implementations for inline forms. We are
introducing `.form-row` and respective variants to different viewports
`.form-row-*` to have more control over spacing for inline forms.

Each input element will consume the necessary space it needs and
wrap if not enough space is available.

The spacing for clearable elements is now applied to
`.form-control-clearable`. This allows the spacing adjustments to be
present before the JavaScript is invoked. The class will be added
through js if not already present on the element. This avoids
additional layout shifts.

Resolves: #100663
Releases: main
Change-Id: If9757d9dc04d9dca83cb1e89c5bbced1b20b5850
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78682
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>