[!!!][TASK] Remove UserTSconfig option "overridePageModule"
The UserTSconfig setting options.overridePageModule
was introduced in TYPO3 v4, to run TemplaVoila
and "regular" page module in parallel. However, this
feature was never rolled out completely throughout
the TYPO3 Backend, but only used in a few places,
and now it is only used in a few places (mainly workspaces).

Since TYPO3 v12, it is much easier to actually override
the page module through the Module API to replace
either the endpoint or to add a new module with an alias
to "web_layout", which should be used instead.

Resolves: #102499
Releases: main
Change-Id: Ibfd4efa221a780f72200364c4e5fd65dc64e3be7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81812
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>