[TASK] Remove wizard component from workspaces management
The wizard component was introduced in EXT:workspaces back then as the
`Modal` component was not able to keep the current window open as long
an action is running.

Since any action executed in the workspaces management module is not
wizard-esque, plain modal windows are used instead now. The usage of
the `Wizard` component is completely dropped.

Resolves: #103160
Releases: main
Change-Id: Icb4b746816f872ec8e5e335823956917a7683ecd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83045
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>