[TASK] Improve usability of search in workspaces
This changes the `disabled`-attribute of the workspaces'
search-submit-button. Thereby screen-readers can announce that the
button is disabled.

Before this there was toggling implemented for `class="disabled"`.
Thereby the 'disabled' button was still active and could be activated
by keyboard.

This also enables a proper reload if the user deletes all characters
from the search-field.
Before on emptying the search-field a reload was already triggered, but
the search-request still used the old non-empty search-string.

This triggers a reload on change of the search-field.
Before a user might have entered a search-string and then changed one
of the dropdowns. Thereby a reload was triggerd, that did not use the
new search-string, even though it was displayed in the search-field.

Resolves: #104885
Releases: main
Change-Id: I3d3660f27ca22127444cbf62489b2ad0efc8cbb5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85997
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>