[BUGFIX] Use backed enums in Workspace comparison
When introducing VersionState Enum
for t3ver_state, some issues were introduced
which PHPStan actually found to be (still) valid.

The main issue happened when deleting a record
in a workspace, which might have impact on translations
of this record in a workspace, which might not be shown
in the Workspaces module anymore.

They are removed along with a few minor
other changes related to PHPStan.

Resolves: #106140
Related: #101175
Releases: main, 13.4
Change-Id: Ibec88ad7d6ebbc7bf4d608a6089063eeabb36aae
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88131
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>