[TASK] Hand StageChangeMessage around in ext:workspaces
The "stage changed message" data object has been established
with #97700. It is easily possible to hand it around instead
of dealing with properties in StageChangeNotificationHandler.

Also, StageChangeNotification does not need to be public
anymore and we use named arguments for the final readonly
StageChangeMessage according to our rules.

Resolves: #106154
Related: #97700
Releases: main, 13.4
Change-Id: I84d1d14d4bcb180f06615466e32ecc6eb3273da2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88170
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>