[BUGFIX] Fallback to placeholder on missing icon in workspace table
If a new root page is created in a workspace, it may lack a site
configuation. This leads to the situation that an icon is attempted to
be considered for the default language, which fails as there is no
configuration, yet.

The code responsible for the failing call is even declared to return
null values, which wasn't detected by phpstan (probably due to a too low
level). In case such language icon cannot be determined, an empty
placeholder icon is now used.

Resolves: #106445
Releases: main, 13.4, 12.4
Change-Id: Ic94ff16e61ca1ca1b289703c1e9d905bb3873d02
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88879
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>