[TASK] Simplify LabelCapability
The LabelCapability used `FieldTypeInterface`
objects instead of just the related field names.

This in the past (#106362) already required to
introduce the `SystemInternalFieldType` to
handle fields, not configured as TCA column,
e.g. `uid`.

Since it has been observed, that basically all
usages of the LabelCapability always access the
`getName()` method of the fields, this patch
now simplifies the LabelCapability by just
providing the configured field names.

There are only two places where the actual field
is evaluated. This is however easily possible
by just fetching the field via the TcaSchema
with `getField()`.

Resolves: #106873
Releases: main, 13.4
Change-Id: I03d88d52db3066c23996fe64be80665fd3f793d6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89713
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>