[TASK] Workspace module action bar is rendered once
The workspace module action bar was rendered at the top and at
the bottom of the record list in TYPO3 v12. Showing the bottom
bar broke with #102431. While this was most likely not intended,
it makes the module more slick, so we keep it. The patch removes
the duplicated but css-hidden rendering of the bottom bar.

Resolves: #106880
Related: #102431
Releases: main, 13.4
Change-Id: Ibaa51dd8ca70eef66e6bbd6e2d81675949523112
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89724
Tested-by: Stefan Froemken <froemken@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>