[BUGFIX] Respect TSconfig disableButtonForDokType everywhere
This change is the first in a series of patches to
streamline preview link generation.

This first patch puts the logic what kind of page preview links
can be generated into one spot: The PreviewUriBuilder.

Thus, the TSconfig option TCEMAIN.preview.disableButtonForDokType
is always respected, and this means that the icon is now
removed in Page module for sysfolders etc.

The PreviewUriBuilder can now receive a array instead
of just the Page ID, reducing the calling code logic
drastically.

Resolves: #96861
Releases: main, 13.4
Change-Id: Ic8aa2006d8a86ea5f7d4850f6ced6a21060a060e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89815
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>