[TASK] Simplify a query in WorkspaceService
It is enough to look for t3ver_wsid = X to find workspace
records in a given table. The additional clause looking
for t3ver_oid > 0 (modified, to-delete, moved) OR
t3ver_state = 1 (new) can be left out since it does not
reduce the result set any further.

Resolves: #106940
Releases: main, 13.4
Change-Id: I282d08f130c4adb9eaba9cb9dab06a2ab101c835
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89835
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>