[TASK] Extract workspace internal table hook code
The main ext:workspaces DataHandlerHook is shrinking
for a while already: We're integrating more and more
into DataHandler directly to enable streamlining and
refactoring. One of the next steps will be looking at
the remaining "publish" and "setStage" code to see
what could be moved to DataHandler as well.

In addition to this, this patch extracts two other
standalone parts of DataHandlerHook to an isolated
class DataHandlerInternalWorkspaceTablesHook: It
handles things that have to be done when
sys_workspace_stage and sys_workspace rows are
editited in the backend. The handling when deleting
a sys_workspace_stage is also slightly improved by
adding a log entry stating affected record counts.

Resolves: #106959
Releases: main, 13.4
Change-Id: Ied8999510cc0c7723f45498acbff69879efe2150
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89854
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>