[TASK] Avoid workspace selector in workspace module
The workspace selector within the docheader of the workspaces
module is a straight duplication of the toolbar workspaces
selector. The patch removes the docheader drop down since it
is not only affecting the workspace module content, but is
a global switch that sets context for the entire backend.

Resolves: #107048
Releases: main, 13.4
Change-Id: I841606004e5a49d8bc0a596b272c7c1e8575b9ef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89998
Tested-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>