[TASK] Simplify workspaces PreviewController
The workspaces "split" preview renders the "initial" stage
change buttons server side, and updates them with an ajax
call when stages are changed or changes are discarded.
This leads to duplicate code: PreviewController and the
ajax call do the same thing. The patch simplifies this by
not delivering available stages up front, but fetches the
buttons via ajax directly after load. The call is quite
quick - much quicker than rendering the page itself, so
it does not harm usability.

Resolves: #107064
Releases: main, 13.4
Change-Id: Ie2d80cc18a7f8d98ae0777ded3ff4338ad004faf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90038
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>