[BUGFIX] Improve workspace stagechg_notification
sys_workspace records have a select drop down with three
options to globally determine if notifications upon stage
changes of workspace records are sent.

There is no difference in its handling between state 1
and 10. The patch turns the select into an on/off toggle
and improves labeling and description.

Note, if disabled, this checkbox technically obsoletes the
'show dialog' checkbox setting available for single stages,
which in turn influences the "Allow to change pre-selection
of recipients" checkbox. We'll improve wording of them with
another change to hint about relations.

Resolves: #76224
Related: #90411
Releases: main, 13.4
Change-Id: Ibc67831ed731201cd35379148094f5577003cba6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90052
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>