[BUGFIX] Split workspace AJAX routes into different scopes
The `workspace_dispatch` route was hardened in #106983, limiting access
to users with access to the workspaces backend module only. It was
however forgotten that the AJAX route is used in workspace preview
context as well to render buttons, discard changes and send changes to
different stages. Since this functionality must be available to all
backend users having the general workspace functionality, the
appropriate workspace actions are now reflected in a new AJAX route
`workspace_preview`. The new AJAX route does not define route access
permissions – however, an additional access check is now performed in
the new dedicated controller action.

Resolves: #107976
Related: #106983
Releases: main, 13.4, 12.4
Change-Id: If6c92ae4d65338900b65cff1663cfd175dade056
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91538
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>